+++ src/lambdaisland/glogi.clj
@@ -25,3 +25,27 @@
 
 (defmacro error [& keyvals]
   (log-expr &form :error keyvals))
+
+;; goog.log specific
+
+(defmacro shout [& keyvals]
+  (log-expr &form :shout keyvals))
+
+(defmacro severe [& keyvals]
+  (log-expr &form :severe keyvals))
+
+(defmacro fine [& keyvals]
+  (log-expr &form :fine keyvals))
+
+(defmacro finer [& keyvals]
+  (log-expr &form :finer keyvals))
+
+(defmacro finest [& keyvals]
+  (log-expr &form :finest keyvals))
+
+(defmacro spy [form]
+  (let [res (gensym)]
+    `(let [~res ~form]
+       ~(log-expr &form :debug [:spy `'~form
+                                :=> res])
+       ~res)))