+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.0-41 (2020-03-31 / ab9f97f)
 
 ## Added